pyproject.toml
[project]
name = "social-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "apscheduler>=3.11.0",
    "google-adk>=1.16.0",
    "google-genai>=1.45.0",
    "openai>=2.5.0",
    "requests>=2.32.5",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "sentence-transformers>=2.2.0",
    "tenacity>=8.0.0",
]
